Organise connector components into subfolders
