+++ src/components/connectors/jolocom/JolocomIssue.vue
