+++ src/components/connectors/jolocom/JolocomVerify.vue
