+++ src/components/connectors/trinsic/TrinsicConnectionModal.vue
