+++ src/components/connectors/trinsic/TrinsicIssue.vue
